[REF] mail: split DiscussCategories in many files
The file defined many components and templates, which while this
reduces the amount of files, this makes it hard to work in this file.

This commit splits in as many files as there are components.
