+++ addons/hr_holidays/static/src/discuss/core/public_web/discuss_sidebar_channel_patch.xml
