+++ addons/mail/static/src/core/public_web/discuss_sidebar.scss
@@ -27,11 +27,17 @@
     }
 }
 
+.o-mail-DiscussSidebar-floatingMenu {
+    max-width: 250px !important;
+}
+
 .o-mail-DiscussSidebar-item {
     outline: 1px solid transparent !important;
     outline-offset: -1px;
     padding-top: map-get($spacers, 1) / 2;
     padding-bottom: map-get($spacers, 1) / 2;
+    line-height: 2; // so that same height as actions
+    background-color: inherit !important;
 
     .o-mail-DiscussSidebar.o-compact & {
         opacity: 75%;
@@ -56,4 +62,9 @@
             filter: brightness(1.1);
         }
     }
+}
+
+.o-mail-DiscussSidebar-unreadIndicator {
+    font-size: 0.4rem;
+    left: 0px;
 }
