+++ addons/mail/static/src/core/public_web/discuss_sidebar.xml
@@ -17,4 +17,8 @@
             </ResizablePanel>
         </div>
     </t>
+
+    <t t-name="mail.DiscussSidebar.unreadIndicator">
+        <span class="o-mail-DiscussSidebar-unreadIndicator position-absolute" t-att-class="{ 'o-compact': store.discuss.isSidebarCompact }" title="Thread has unread messages"><i class="fa fa-circle smaller"/></span>
+    </t>
 </templates>
