+++ addons/mail/static/src/discuss/call/public_web/discuss_sidebar_call_indicator.js
@@ -1,5 +1,5 @@
 import { Thread } from "@mail/core/common/thread_model";
-import { discussSidebarChannelIndicatorsRegistry } from "@mail/discuss/core/public_web/discuss_sidebar_categories";
+import { discussSidebarChannelIndicatorsRegistry } from "@mail/discuss/core/public_web/discuss_sidebar_channel";
 
 import { Component } from "@odoo/owl";
 import { useService } from "@web/core/utils/hooks";
