+++ addons/mail/static/src/discuss/call/public_web/discuss_sidebar_channel_patch.dark.scss
