+++ addons/mail/static/src/discuss/core/public_web/discuss_sidebar_channel.dark.scss
@@ -1,7 +1,3 @@
-.o-mail-DiscussSidebar-item {
-    --mail-DiscussSidebarChannel-commandsHover: #{rgba($gray-400, .5)};
-}
-
 .o-mail-DiscussSidebarChannel-container {
     --mail-DiscussSidebarChannel-borderOpacity: .125;
     ---mail-DiscussSidebarChannel-borderedBgColor: #{mix($gray-100, $gray-200, 75%)};
@@ -9,8 +5,4 @@
 
 .o-mail-DiscussSidebarChannel-floatingName {
     ---mail-DiscussSidebarChannel-floatingNameBgColor: #{mix($white, $gray-100, 25%)};
-}
-
-.o-mail-DiscussSidebarSubchannel {
-    --mail-DiscussSidebarSubchannel-svgColor: #{mix($gray-300, $gray-400)};
 }
