+++ addons/mail/static/src/discuss/core/public_web/discuss_sidebar_channel.scss
@@ -1,0 +1,63 @@
+.o-mail-DiscussSidebarChannel-actions {
+    .o-mail-DiscussSidebar-item:hover & {
+        display: flex !important;
+    }
+
+    button {
+        --border-opacity: .15;
+    
+        &.rounded-circle {
+            --oi-vertical-align: -5%;
+            aspect-ratio: 1;
+        }
+    
+        @media (min-resolution: 2dppx) {
+            .oi-ellipsis-h {
+                transform: translateX(0.5px);
+            }
+        }
+    
+        &:hover, &:active, &:focus-visible, &.o-showingActions {
+            --border-opacity: 1;
+            --border-color: #{$o-action};
+        }
+    }
+}
+
+.o-mail-DiscussSidebarChannel-badge.o-nonCompact {
+    margin-right: map-get($spacers, 2) + map-get($spacers, 2) / 2;
+}
+
+.o-mail-DiscussSidebarChannel-container {
+    --border-opacity: var(--mail-DiscussSidebarChannel-borderOpacity, .05);
+    gap: 1px;
+
+    &.o-bordered {
+        background-color: var(---mail-DiscussSidebarChannel-borderedBgColor, mix($white, $dark, 98%)) !important;
+
+        .o-mail-DiscussSidebar-unreadIndicator {
+            left: -1px; // compensate for border
+
+            &.o-mail-DiscussSidebarChannel, &.o-compact {
+                filter: drop-shadow(1px 2px 0px $white) drop-shadow(-1px 1px 0px $white) drop-shadow(1px -1px 0px $white) drop-shadow(-1px -2px 0px $white);
+            }
+        }
+    }
+}
+
+.o-mail-DiscussSidebarChannel-itemMain {
+    background-color: inherit !important; // Not 'bg-inherit' so it cancels btn hover effect too
+}
+
+.o-mail-DiscussSidebarChannel-indicatorCompact {
+    top: 1px;
+    left: 3px;
+}
+
+.o-mail-DiscussSidebarChannel-subChannel.o-nonCompact {
+    margin-left: map-get($spacers, 4) + map-get($spacers, 2);
+}
+
+.o-mail-DiscussSidebarChannel-subChannelList {
+    gap: 1px;
+}
