+++ addons/mail/static/src/discuss/core/public_web/discuss_sidebar_subchannel.dark.scss
@@ -1,0 +1,3 @@
+.o-mail-DiscussSidebarSubchannel {
+    --mail-DiscussSidebarSubchannel-svgColor: #{mix($gray-300, $gray-400)};
+}
